requirements.txt
numpy
scipy
torch
torchvision
tensorboard
ray
pandas
metayaml
oyaml
filelock
humanfriendly
tqdm
glom
deepmerge
pathspec
imageio
imageio-ffmpeg
imagecodecs
inflection
recordtype
matplotlib
seaborn
pygame
gym
psutil
msgpack
msgpack-numpy
pyarrow
py-lz4framed
pyzmq
tenacity
toolz
port-for
Jinja2<3
